feat: consistent theme & cleanup for auth widgets
