+++ designer_v2/lib/features/auth/password_forgot_page.dart
@@ -5,7 +5,7 @@ import 'package:studyu_designer_v2/common_views/primary_button.dart';
 import 'package:studyu_designer_v2/features/auth/auth_controller.dart';
 import 'package:studyu_designer_v2/features/auth/auth_state.dart';
 import 'package:studyu_designer_v2/features/auth/form_controller.dart';
-import 'package:studyu_designer_v2/features/auth/form_widgets.dart';
+import 'package:studyu_designer_v2/features/auth/auth_formfield_views.dart';
 import 'package:studyu_designer_v2/localization/string_hardcoded.dart';
 import 'package:studyu_designer_v2/routing/router.dart';
 import 'package:studyu_designer_v2/routing/router_intent.dart';
@@ -57,7 +57,7 @@ class _PasswordForgotPageContentState extends ConsumerState<PasswordForgotPageCo
               child: Text('Forgot Password'.hardcoded, /*style: FlutterFlowTheme.of(context).title1,*/)
           ),
           const SizedBox(height: 20),
-          const TextFormFieldWidget(),
+          const EmailTextField(),
           const SizedBox(height: 20),
           ReactiveFormConsumer(
             builder: (context, formN, child) {
